:hammer: BE: Hotfix - Perubahan Tipe Data Harga Menjadi INT

+++ server/models/Ruangan.js
@@ -57,7 +57,7 @@ const Ruangan = db.define('Ruangan', {
         allowNull: true
     },
     harga_barang: {
-        type: Sequelize.REAL,
+        type: Sequelize.INTEGER,
         allowNull: true
     },
     keadaan_barang: {
